Fix deprecation warning
Before this change running `bundle exec rake spec` resulted in the following warning:

Deprecation Warnings:
puppetlabs_spec_helper: defaults `mock_with` to `:mocha`.

The config option *cannot* be added to an existing block, it has
to be in a seperate block before the gem is loaded to remove
the deprecation.

https://stackoverflow.com/questions/54059220/confusing-rspec-puppet-deprecation-warning-defaults-mock-with-to-mocha
https://tickets.puppetlabs.com/browse/PDK-916